deployment/backend/requirements.txt
waitress
pandas
flask
scikit-learn
tensorflow-cpu
numpy
feature_engine